requirements.txt
duckdb>=1.1.3
numpy>=1.26.4
pandas>=2.2.3
plotly>=5.24.1
python-dateutil>=2.9.0.post0
requests>=2.32.3
PyYAML>=6.0.2
scikit-learn>=1.5.2
scipy>=1.14.1
statsmodels>=0.14.4
streamlit>=1.40.0
tqdm>=4.67.1
